% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{WSC Dataset}
\format{
A data frame with X rows and Y variables:
\describe{
\item{female}{Responder's gender is female}
\item{white}{Responder's race is White}
...
\item{logBDI}{log of BDI scale }

}
}
\source{
\url{https://example.com}

{B.Keller et al. (2024)}....
}
\usage{
my_data
}
\description{
A dataset containing 2200 observations and 44 columbns used for WSC four arm study.
}
\examples{
# example code




}
\keyword{datasets}
