% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{WSC}
\format{
A data frame with X rows and Y variables:
\describe{
\item{var1}{Description of var1}
\item{var2}{Description of var2}

}
}
\source{
\url{https://example.com}
}
\usage{
my_data
}
\description{
A dataset containing XYZ information.
}
\keyword{datasets}
